tercera practica integradora completa

+++ src/controllers/product.controller.js
@@ -41,6 +41,8 @@ class ProductController{
     createProduct = async(req, res, next)=>{
         try{
             const {title, description, price, code, stock, category, thumbnail} = req.body
+            const user = req.session.user
+
             if(!title || !description || !price || !code || !stock || !category){
                 CustomError.createError({
                     name: 'Product creation error',
@@ -56,7 +58,13 @@ class ProductController{
                     code: Error.INVALID_TYPE_ERROR
                 })
             }
-            let newProduct = new ProductDto({title, description, price, code, stock, category, thumbnail}) 
+
+            let owner = 'admin'
+            if(user && user.role === 'premium'){
+                owner = user.email
+            }
+
+            let newProduct = new ProductDto({title, description, price, code, stock, category, thumbnail, owner}) 
             let product = await productService.createProduct(newProduct)
             !product
             ? res.status(400).send({ error: "No se pudo crear el producto" })
@@ -81,11 +89,19 @@ class ProductController{
 
     deleteProduct = async(req, res)=>{
         try{
-            const id = req.params.pid;
-            const deletedProduct = await productService.deleteProduct(id)
-            !deletedProduct
-            ? res.status(404).send({error: `El producto con ID ${id} no existe`})
-            : res.status(200).send({ status:`El producto con ID ${id} se ha eliminado`});
+            const id = req.params.pid
+            const user = req.session.user
+
+            const product = await productService.getProductById(id)
+            if (!product) return res.status(404).send({ error: `El producto con ID ${id} no existe` })
+            
+            if(user && (user.role === 'admin' || (user.role === 'premium' && product.owner === user.email))){
+                const deletedProduct = await productService.deleteProduct(id)
+                if (deletedProduct) {
+                    return res.status(200).send({ status: `El producto con ID ${id} se ha eliminado` });
+                }
+            }
+            return res.status(403).send({ error: "No tienes permiso para eliminar este producto" })
         }catch(error){
             logger.error(error)
         }
